fix: move errors to pkg
