+++ internal/subscribers/handler_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/gorilla/mux"
-	utils "github.com/skobelina/currency_converter/pkg/utils/errors"
+	utils "github.com/skobelina/currency_converter/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
